pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "apm-cli"
version = "0.0.14"
description = "MCP configuration tool"
readme = "README.md"
requires-python = ">=3.9"
license = {file = "LICENSE"}
authors = [
    {name = "Daniel Meppiel", email = "user@example.com"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "click>=8.0.0",
    "colorama>=0.4.6",
    "pyyaml>=6.0.0",
    "requests>=2.28.0",
    "python-frontmatter>=1.0.0",
    "llm>=0.17.0",
    "llm-github-models>=0.1.0",
    "tomli>=1.2.0; python_version<'3.11'",
    "rich>=13.0.0",
    "rich-click>=1.7.0",
    "watchdog>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "mypy>=1.0.0",
]
build = [
    "pyinstaller>=6.0.0",
]

[project.scripts]
apm = "apm_cli.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["apm_cli*"]

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
